Bugfix: support a choice of pubkey for encryption.
The program was intended to support multiple encryption keys but did
not. The intended functionality seemed to be as follows:
 - Have the user input a GPG ID
 - Extract a list of public encryption keys associated with that GPG ID.
 - Encrypt with all those pubkeys as recipients.
In reality, what happened was that the following gpg command was run:
gpg -e -r pubkey1 -r pubkey2 -r pubkey3 ...
which actually runs through each key, and chooses a default choice of
pubkey belonging to the identity which owns each given key. So we end up
encrypting actually for just one key. See here for more info:
https://security.stackexchange.com/a/181640/142660

This has been fixed with the following changes. From a design
perspective:
First, the user is no longer encouraged to input a single GPG identity
with `pass init gpg-id` but instead a list of public encryption keys as
per `pass init gpg-pubkey1 [gpg-pubkey2 [gpg-pubkey3 ..]]`. The
resulting behaviour is easier to predict, more precise and easier to
implement.

The actual changes made to code are here summarised:
 - Add exclamation mark after each GPG key in `gpg -e -r pubkeyX! ...`
 - The program displayed false information about the keys for which
   files were being encryted. This is fixed by replacing
   `${gpg_keys//$'\n'/ }` with `${GPG_RECIPIENTS[@]}`.
 - Function help messages did not show that multiple keys were
   supported. I have changed the help messages in the style of:
   `pass init gpg-id`
   --> `pass init gpg-pubkey1 [gpg-pubkey2 [gpg-pubkey3 ..]]`

The filename `.gpg-id` has been left unchanged.